refactor: move pure function to separate file
